base sqlalchemy classes
