+++ simpleml/orm/metadata.py
@@ -1,62 +1,21 @@
-"""
-Base class for sqlalchemy
-"""
+'''
+Metadata bases specific to each session (subclasses represent tables affected by the
+same session -- ie base.metadata.create_all()/drop_all()/upgrade())
+
+Each class used as part of a session needs to be initialized directly
+'''
 
 __author__ = "Elisha Yadgaran"
 
+
 import logging
 
-from sqlalchemy import DDL, Column, DateTime, MetaData, event, func
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy_mixins import AllFeaturesMixin
+from sqlalchemy import event, MetaData, DDL
 
-Base = declarative_base()
+from .base_sqlalchemy import BaseSQLAlchemy
+
+
 LOGGER = logging.getLogger(__name__)
-
-
-class BaseSQLAlchemy(Base, AllFeaturesMixin):
-    """
-    Base class for all SimpleML database objects. Defaults to PostgreSQL
-    but can be swapped out for any supported SQLAlchemy backend.
-
-    Takes advantage of sqlalchemy-mixins to enable active record operations
-    (TableModel.save(), create(), where(), destroy())
-
-    Added some inheritable convenience methods
-
-    -------
-    Schema
-    -------
-    created_timestamp: Server time on insert
-    modified_timestamp: Server time on update
-    """
-
-    __abstract__ = True
-
-    created_timestamp = Column(
-        DateTime(timezone=True), nullable=False, server_default=func.now()
-    )
-    modified_timestamp = Column(DateTime(timezone=True), server_onupdate=func.now())
-
-    @classmethod
-    def filter(cls, *filters):
-        return cls._session.query(cls).filter(*filters)
-
-    @classmethod
-    def query_by(cls, *queries):
-        return cls._session.query(*queries)
-
-
-@event.listens_for(BaseSQLAlchemy, "before_update", propagate=True)
-def _receive_before_update(mapper, connection, target):
-    """Listen for updates and update `modified_timestamp` column."""
-    target.modified_timestamp = func.now()
-
-
-"""
-Metadata bases specific to each session (subclasses represent tables affected by the
-same session -- ie base.metadata.create_all()/drop_all()/upgrade())
-"""
 
 
 class SimplemlCoreSqlalchemy(BaseSQLAlchemy):
